app/requirements.txt
streamlit==1.41.1
tensorflow==2.12.0
pillow
numpy
